+++ tests/validations/test_metrics.py
@@ -1,4 +1,3 @@
-from copy import deepcopy
 from typing import List, Tuple
 
 import pytest
@@ -31,7 +30,6 @@ from dbt_semantic_interfaces.references import (
     TimeDimensionReference,
 )
 from dbt_semantic_interfaces.test_utils import (
-    find_metric_with,
     metric_with_guaranteed_meta,
     semantic_model_with_guaranteed_meta,
 )
@@ -48,7 +46,6 @@ from dbt_semantic_interfaces.validations.metrics import (
     CumulativeMetricRule,
     DerivedMetricRule,
     MetricTimeGranularityRule,
-    WhereFiltersAreParseable,
 )
 from dbt_semantic_interfaces.validations.semantic_manifest_validator import (
     SemanticManifestValidator,
@@ -343,145 +340,6 @@ def test_derived_metric() -> None:  # noqa: D
         "No `expr` set for derived metric",
     ]
     check_error_in_issues(error_substrings=expected_substrings, issues=build_issues)
-
-
-def test_where_filter_validations_happy(  # noqa: D
-    simple_semantic_manifest__with_primary_transforms: PydanticSemanticManifest,
-) -> None:
-    validator = SemanticManifestValidator[PydanticSemanticManifest]([WhereFiltersAreParseable()])
-    results = validator.validate_semantic_manifest(simple_semantic_manifest__with_primary_transforms)
-    assert not results.has_blocking_issues
-
-
-def test_where_filter_validations_bad_base_filter(  # noqa: D
-    simple_semantic_manifest__with_primary_transforms: PydanticSemanticManifest,
-) -> None:
-    manifest = deepcopy(simple_semantic_manifest__with_primary_transforms)
-
-    metric, _ = find_metric_with(manifest, lambda metric: metric.filter is not None)
-    assert metric.filter is not None
-    assert len(metric.filter.where_filters) > 0
-    metric.filter.where_filters[0].where_sql_template = "{{ dimension('too', 'many', 'variables', 'to', 'handle') }}"
-    validator = SemanticManifestValidator[PydanticSemanticManifest]([WhereFiltersAreParseable()])
-    with pytest.raises(SemanticManifestValidationException, match=f"trying to parse filter of metric `{metric.name}`"):
-        validator.checked_validations(manifest)
-
-
-def test_where_filter_validations_bad_measure_filter(  # noqa: D
-    simple_semantic_manifest__with_primary_transforms: PydanticSemanticManifest,
-) -> None:
-    manifest = deepcopy(simple_semantic_manifest__with_primary_transforms)
-
-    metric, _ = find_metric_with(
-        manifest, lambda metric: metric.type_params is not None and metric.type_params.measure is not None
-    )
-    assert metric.type_params.measure is not None
-    metric.type_params.measure.filter = PydanticWhereFilterIntersection(
-        where_filters=[
-            PydanticWhereFilter(where_sql_template="{{ dimension('too', 'many', 'variables', 'to', 'handle') }}")
-        ]
-    )
-    validator = SemanticManifestValidator[PydanticSemanticManifest]([WhereFiltersAreParseable()])
-    with pytest.raises(
-        SemanticManifestValidationException,
-        match=f"trying to parse filter of measure input `{metric.type_params.measure.name}` on metric `{metric.name}`",
-    ):
-        validator.checked_validations(manifest)
-
-
-def test_where_filter_validations_bad_numerator_filter(  # noqa: D
-    simple_semantic_manifest__with_primary_transforms: PydanticSemanticManifest,
-) -> None:
-    manifest = deepcopy(simple_semantic_manifest__with_primary_transforms)
-
-    metric, _ = find_metric_with(
-        manifest, lambda metric: metric.type_params is not None and metric.type_params.numerator is not None
-    )
-    assert metric.type_params.numerator is not None
-    metric.type_params.numerator.filter = PydanticWhereFilterIntersection(
-        where_filters=[
-            PydanticWhereFilter(where_sql_template="{{ dimension('too', 'many', 'variables', 'to', 'handle') }}")
-        ]
-    )
-    validator = SemanticManifestValidator[PydanticSemanticManifest]([WhereFiltersAreParseable()])
-    with pytest.raises(
-        SemanticManifestValidationException, match=f"trying to parse the numerator filter on metric `{metric.name}`"
-    ):
-        validator.checked_validations(manifest)
-
-
-def test_where_filter_validations_bad_denominator_filter(  # noqa: D
-    simple_semantic_manifest__with_primary_transforms: PydanticSemanticManifest,
-) -> None:
-    manifest = deepcopy(simple_semantic_manifest__with_primary_transforms)
-
-    metric, _ = find_metric_with(
-        manifest, lambda metric: metric.type_params is not None and metric.type_params.denominator is not None
-    )
-    assert metric.type_params.denominator is not None
-    metric.type_params.denominator.filter = PydanticWhereFilterIntersection(
-        where_filters=[
-            PydanticWhereFilter(where_sql_template="{{ dimension('too', 'many', 'variables', 'to', 'handle') }}")
-        ]
-    )
-    validator = SemanticManifestValidator[PydanticSemanticManifest]([WhereFiltersAreParseable()])
-    with pytest.raises(
-        SemanticManifestValidationException, match=f"trying to parse the denominator filter on metric `{metric.name}`"
-    ):
-        validator.checked_validations(manifest)
-
-
-def test_where_filter_validations_bad_input_metric_filter(  # noqa: D
-    simple_semantic_manifest__with_primary_transforms: PydanticSemanticManifest,
-) -> None:
-    manifest = deepcopy(simple_semantic_manifest__with_primary_transforms)
-
-    metric, _ = find_metric_with(
-        manifest,
-        lambda metric: metric.type_params is not None
-        and metric.type_params.metrics is not None
-        and len(metric.type_params.metrics) > 0,
-    )
-    assert metric.type_params.metrics is not None
-    input_metric = metric.type_params.metrics[0]
-    input_metric.filter = PydanticWhereFilterIntersection(
-        where_filters=[
-            PydanticWhereFilter(where_sql_template="{{ dimension('too', 'many', 'variables', 'to', 'handle') }}")
-        ]
-    )
-    validator = SemanticManifestValidator[PydanticSemanticManifest]([WhereFiltersAreParseable()])
-    with pytest.raises(
-        SemanticManifestValidationException,
-        match=f"trying to parse filter for input metric `{input_metric.name}` on metric `{metric.name}`",
-    ):
-        validator.checked_validations(manifest)
-
-
-def test_where_filter_validations_invalid_granularity(  # noqa: D
-    simple_semantic_manifest__with_primary_transforms: PydanticSemanticManifest,
-) -> None:
-    manifest = deepcopy(simple_semantic_manifest__with_primary_transforms)
-
-    metric, _ = find_metric_with(
-        manifest,
-        lambda metric: metric.type_params is not None
-        and metric.type_params.metrics is not None
-        and len(metric.type_params.metrics) > 0,
-    )
-    assert metric.type_params.metrics is not None
-    input_metric = metric.type_params.metrics[0]
-    input_metric.filter = PydanticWhereFilterIntersection(
-        where_filters=[
-            PydanticWhereFilter(where_sql_template="{{ TimeDimension('metric_time', 'cool') }}"),
-            PydanticWhereFilter(where_sql_template="{{ TimeDimension('metric_time', 'month') }}"),
-            PydanticWhereFilter(where_sql_template="{{ TimeDimension('metric_time', 'MONTH') }}"),
-        ]
-    )
-    validator = SemanticManifestValidator[PydanticSemanticManifest]([WhereFiltersAreParseable()])
-    issues = validator.validate_semantic_manifest(manifest)
-    assert not issues.has_blocking_issues
-    assert len(issues.warnings) == 1
-    assert "`cool` is not a valid granularity name" in issues.warnings[0].message
 
 
 def test_conversion_metrics() -> None:  # noqa: D
